getting rid of warnings
git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder/trunk/moses@854 1f5c12ca-751b-0410-a591-d2e778427230

+++ src/StaticData.h
@@ -90,7 +90,7 @@ protected:
 	bool m_sourceStartPosMattersForRecombination;
 
 	int m_inputType;
-	unsigned m_numInputScores;
+	size_t m_numInputScores;
 
 	size_t m_verboseLevel;
 	DistortionScoreProducer *m_distortionScoreProducer;
